Move reindex function from memex to h
We will need to keep track of the new index name during the reindex,
which is happening in h.
