+++ tests/memex/search/index_test.py
@@ -81,99 +81,6 @@ class TestDeleteAnnotation:
         )
 
 
-@pytest.mark.usefixtures('BatchIndexer',
-                         'configure_index',
-                         'get_aliased_index',
-                         'update_aliased_index')
-class TestReindex(object):
-    def test_sets_op_type_to_create(self, es, BatchIndexer):
-        index.reindex(mock.sentinel.session, es, mock.sentinel.request)
-
-        _, kwargs = BatchIndexer.call_args
-        assert kwargs['op_type'] == 'create'
-
-    def test_indexes_annotations(self, es, indexer):
-        """Should call .index() on the batch indexer instance."""
-        index.reindex(mock.sentinel.session, es, mock.sentinel.request)
-
-        indexer.index.assert_called_once_with()
-
-    def test_retries_failed_annotations(self, es, indexer):
-        """Should call .index() a second time with any failed annotation IDs."""
-        indexer.index.return_value = ['abc123', 'def456']
-
-        index.reindex(mock.sentinel.session, es, mock.sentinel.request)
-
-        assert indexer.index.mock_calls == [
-            mock.call(),
-            mock.call(['abc123', 'def456']),
-        ]
-
-    def test_creates_new_index(self, es, configure_index, matchers):
-        """Creates a new target index."""
-        index.reindex(mock.sentinel.session, es, mock.sentinel.request)
-
-        configure_index.assert_called_once_with(es)
-
-    def test_passes_new_index_to_indexer(self, es, configure_index, BatchIndexer):
-        """Pass the name of the new index as target_index to indexer."""
-        configure_index.return_value = 'hypothesis-abcd1234'
-
-        index.reindex(mock.sentinel.session, es, mock.sentinel.request)
-
-        _, kwargs = BatchIndexer.call_args
-        assert kwargs['target_index'] == 'hypothesis-abcd1234'
-
-    def test_updates_alias_when_reindexed(self, es, configure_index, update_aliased_index):
-        """Call update_aliased_index on the client with the new index name."""
-        configure_index.return_value = 'hypothesis-abcd1234'
-
-        index.reindex(mock.sentinel.session, es, mock.sentinel.request)
-
-        update_aliased_index.assert_called_once_with(es, 'hypothesis-abcd1234')
-
-    def test_does_not_update_alias_if_indexing_fails(self, es, indexer, update_aliased_index):
-        """Don't call update_aliased_index if index() fails..."""
-        indexer.index.side_effect = RuntimeError('fail')
-
-        try:
-            index.reindex(mock.sentinel.session, es, mock.sentinel.request)
-        except RuntimeError:
-            pass
-
-        assert not update_aliased_index.called
-
-    def test_raises_if_index_not_aliased(self, es, get_aliased_index):
-        get_aliased_index.return_value = None
-
-        with pytest.raises(RuntimeError):
-            index.reindex(mock.sentinel.session, es, mock.sentinel.request)
-
-    @pytest.fixture
-    def BatchIndexer(self, patch):
-        return patch('memex.search.index.BatchIndexer')
-
-    @pytest.fixture
-    def configure_index(self, patch):
-        return patch('memex.search.index.configure_index')
-
-    @pytest.fixture
-    def get_aliased_index(self, patch):
-        func = patch('memex.search.index.get_aliased_index')
-        func.return_value = 'foobar'
-        return func
-
-    @pytest.fixture
-    def update_aliased_index(self, patch):
-        return patch('memex.search.index.update_aliased_index')
-
-    @pytest.fixture
-    def indexer(self, BatchIndexer):
-        indexer = BatchIndexer.return_value
-        indexer.index.return_value = []
-        return indexer
-
-
 class TestBatchIndexer(object):
     def test_index_indexes_all_annotations_to_es(self, db_session, indexer, matchers, streaming_bulk, factories):
         ann_1, ann_2 = factories.Annotation(), factories.Annotation()
